add explicit ratio_of_defaults_for_sparse_serialization type cast

+++ src/Storages/MergeTree/MutateTask.cpp
@@ -706,7 +706,7 @@ getColumnsForNewDataPart(
 
     SerializationInfo::Settings source_part_serialization_settings = SerializationInfo::Settings
     {
-        (*source_part->storage.getSettings())[MergeTreeSetting::ratio_of_defaults_for_sparse_serialization],
+        static_cast<double>((*source_part->storage.getSettings())[MergeTreeSetting::ratio_of_defaults_for_sparse_serialization]),
         false,
         serialization_infos.getSettings().version,
         serialization_infos.getSettings().string_serialization_version,
@@ -716,7 +716,7 @@ getColumnsForNewDataPart(
     };
     SerializationInfo::Settings storage_serialization_settings = SerializationInfo::Settings
     {
-        (*source_part->storage.getSettings())[MergeTreeSetting::ratio_of_defaults_for_sparse_serialization],
+        static_cast<double>((*source_part->storage.getSettings())[MergeTreeSetting::ratio_of_defaults_for_sparse_serialization]),
         false,
         (*source_part->storage.getSettings())[MergeTreeSetting::serialization_info_version],
         (*source_part->storage.getSettings())[MergeTreeSetting::string_serialization_version],
